requirements.txt
tensorflow==2.17.1
opencv-python
numpy
pandas
matplotlib
scikit-learn
streamlit
pillow
kaggle
streamlit-webrtc
av